pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "skillforge"
version = "0.1.0"
description = "云端自动化 AI Agent 平台 - 支持多后端定时执行 Skills"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
keywords = ["ai", "automation", "github-actions", "claude", "gemini"]
authors = [{ name = "SkillForge Team" }]

dependencies = [
    # AI Backends
    "anthropic>=0.40.0",
    "google-generativeai>=0.8.0",
    "openai>=1.50.0",
    
    # Database
    "supabase>=2.0.0",
    
    # Web & Reports
    "jinja2>=3.1.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    
    # Utils
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "typer>=0.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.6.0",
]

[project.scripts]
skillforge = "src.runner:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
